EXSWHTEC-74 - Implement tests for hipMemcpyWithStream
